+++ models/profile.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer
+from sqlalchemy import Column, String, BigInteger
 from database.model import AppBase
 
 class Profile(AppBase):
@@ -6,6 +6,6 @@ class Profile(AppBase):
 
     id = Column(String(20), primary_key=True)
     name = Column(String(50))
-    _timestamp_ = Column('timestamp', Integer)
+    _timestamp_ = Column('timestamp', BigInteger)
     _signature_ = Column(String(260))
